requirements.txt
python-dotenv==0.13.0
pytube3==9.6.4
typing-extensions==3.7.4.2
